requirements.txt
flask
flask_cors
pandas
numpy
joblib
spacy
psycopg2
python-dotenv
scikit-learn
gunicorn
uuid